+++ package.json
@@ -18,7 +18,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.7",
-    "react-router": "^6.12.1"
+    "react-router-dom": "^6.12.1"
   },
   "devDependencies": {
     "@types/react": "^18.0.37",
